refactor: delete unsued files and recompose rest
