--- components/Dashboard/Consumed/Consumed.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Consumed = () => {
-  return (
-    <div className="col-span-1 h-full">
-      <div className="bg-yellow-500 p-4 rounded-3xl h-full "></div>
-    </div>
-  );
-};
-
-export default Consumed;
