--- components/Dashboard/Dashboard.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Dashboard = () => {
-  return <div className="w-full h-full bg-[#f8f8ff] ">
-    dashboard
-  </div>;
-};
-
-export default Dashboard;
