--- components/Dashboard/Total/Total.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Total = () => {
-  return (
-    <div className=" h-1/2">
-    <div className="bg-blue-500 p-4 rounded-3xl h-full"></div>
-  </div>
-  )
-}
-
-export default Total
